% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{as.Date2}
\alias{as.Date2}
\title{Enhanced Date Conversion}
\usage{
as.Date2(x)
}
\arguments{
\item{x}{Character string representing a date}
}
\value{
A Date object
}
\description{
Converts strings to Date objects with multiple format support
}
\details{
Enhanced Date Conversion

Attempts to parse dates in the following formats:
  - YYYY-MM-DD
  - MM/DD/YYYY
  - YYYY/MM/DD
  - Month DD,YYYY
}
\examples{
as.Date2("2024-02-05")
as.Date2("02/05/2024")
}
