% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{bond.convexity}
\alias{bond.convexity}
\title{Calculate Bond Convexity}
\usage{
bond.convexity(
  settle,
  mature,
  coupon,
  freq = 2,
  yield,
  convention,
  comp.freq = freq
)
}
\arguments{
\item{settle}{Settlement date}

\item{mature}{Maturity date}

\item{coupon}{Annual coupon rate (as decimal)}

\item{freq}{Payment frequency per year (default = 2 for semi-annual)}

\item{yield}{Yield to maturity}

\item{convention}{Day count convention}

\item{comp.freq}{Compounding frequency (defaults to payment frequency)}
}
\value{
Convexity measure of the bond
}
\description{
Calculates the convexity of a bond using total cash flow analysis
}
\details{
Calculate Bond Convexity
}
\examples{
bond.convexity(
  settle = as.Date("2024-02-05"),
  mature = as.Date("2029-02-05"),
  coupon = 0.05,
  yield = 0.06,
  freq = 2,
  convention = "30/360"
)
}
