% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{DATE}
\alias{DATE}
\title{Convert to Date Object}
\usage{
DATE(yyyy, mm, dd)
}
\arguments{
\item{yyyy}{Year (numeric)}

\item{mm}{Month (numeric)}

\item{dd}{Day (numeric)}
}
\value{
A Date object
}
\description{
Creates a Date object from separate year, month, and day values
}
\details{
Convert Year, Month, Day to Date
}
\examples{
DATE(2024, 2, 5)
}
