% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{gfun4}
\alias{gfun4}
\title{Generic Function for Yield Curve Fitting}
\usage{
gfun4(ttm, parm)
}
\arguments{
\item{ttm}{Time to maturity}

\item{parm}{Vector of 6 parameters [a0, a1, a2, a3, a4, a5]}
}
\value{
Fitted yield value
}
\description{
Implements a 5th degree polynomial function for yield curve fitting
}
\details{
Generic Function for Yield Curve Fitting

Uses the formula: a0 + a1*ln(1+t) + a2*ln(1+t)^2 + a3*ln(1+t)^3 + a4*ln(1+t)^4 + a5*ln(1+t)^5
}
\examples{
parms <- c(0.05, 0.02, 0.01, 0.005, 0.002, 0.001)
gfun4(2, parms)
}
