% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{zcb.price}
\alias{zcb.price}
\title{Calculate Zero-Coupon Bond Price}
\usage{
zcb.price(zcb.yield, ttm, freq = 2)
}
\arguments{
\item{zcb.yield}{Yield rate (decimal)}

\item{ttm}{Time to maturity in years}

\item{freq}{Payment frequency per year (default = 2 for semi-annual)}
}
\value{
Price of the zero-coupon bond (per 100 par value)
}
\description{
Calculates the price of a zero-coupon bond given yield, time to maturity, and payment frequency
}
\details{
Zero-Coupon Bond Price Calculator
}
\examples{
zcb.price(0.05, 2, 2)  # 5\% yield, 2 years to maturity, semi-annual compounding
}
