% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Professor_functions.R
\name{zcb.yield}
\alias{zcb.yield}
\title{Calculate Zero-Coupon Bond Yield}
\usage{
zcb.yield(zcb.price, ttm, freq = 2)
}
\arguments{
\item{zcb.price}{Price of the zero-coupon bond}

\item{ttm}{Time to maturity in years}

\item{freq}{Payment frequency per year (default = 2 for semi-annual)}
}
\value{
Yield of the zero-coupon bond (as decimal)
}
\description{
Calculates the yield of a zero-coupon bond given price, time to maturity, and payment frequency
}
\details{
Zero-Coupon Bond Yield Calculator
}
\examples{
zcb.yield(95, 2, 2)  # Price of 95, 2 years to maturity, semi-annual compounding
}
